refactor(actor): extract motivation proposal helpers into a registry skeleton
Move move/patrol/random proposal builders into proposal-helpers.js and add a
per-kind motivations/ module registry. Behaviour is unchanged — dispatch still
lives in the actor controller; modules exist so later milestones can route
through them without an import cycle.
