+++ packages/runtime/src/personas/actor/controller.js
@@ -1,7 +1,6 @@
 import { createActorStateMachine, ActorStates } from "./state-machine.js";
 import { TickPhases } from "../_shared/tick-state-machine.mts";
 import { buildAction, buildRequestActionsFromEffects, buildSolverRequestEffect } from "../_shared/persona-helpers.mts";
-import { EIGHT_WAY_DELTAS } from "../_shared/movement-directions.js";
 import { requireClock } from "../_shared/require-clock.js";
 import {
   RUNTIME_DECISION_CONTRACT,
@@ -28,6 +27,24 @@ import {
   resolveExposureVitalDeltas,
 } from "../../../../core-ts/src/index.ts";
 import { VitalKind } from "../../../../core-ts/src/state/vitals.ts";
+import {
+  DEFAULT_DELTAS,
+  buildMoveProposal,
+  buildPatrolProposals,
+  buildRandomMoveProposals,
+  resolveObservationView,
+  resolveBaseTiles,
+  resolveExit,
+  resolveActor,
+  resolveActorRecord,
+  resolveTileKinds,
+  buildAdjacentMoveProposals,
+  chebyshevDistance,
+  isPassable,
+  findPath,
+  isOccupied,
+  isReserved,
+} from "./proposal-helpers.js";
 
 /**
  * Stage B — `profileAlignment` split into its two independent signals.
@@ -303,7 +320,6 @@ export const actorSubscribePhases = Object.freeze([TickPhases.OBSERVE, TickPhase
 
 const SOLVER_ENGINE = "z3";
 
-const DEFAULT_DELTAS = EIGHT_WAY_DELTAS;
 
 const MOTIVATED_KIND = 2;
 
@@ -356,19 +372,6 @@ function isMotivatedActor(actorId, view, observation) {
   return isMotivatedKind(kind);
 }
 
-function findExitFromTiles(baseTiles) {
-  if (!Array.isArray(baseTiles)) {
-    return null;
-  }
-  for (let y = 0; y < baseTiles.length; y += 1) {
-    const row = String(baseTiles[y]);
-    const x = row.indexOf("E");
-    if (x !== -1) {
-      return { x, y };
-    }
-  }
-  return null;
-}
 
 // CR.6 — resolves from the payload ONLY. This used to fall back to a
 // `lastObservation` cached in the persona closure, which meant a propose could be
@@ -382,71 +385,12 @@ function resolveObservation(payload) {
   return null;
 }
 
-function resolveObservationView(observation) {
-  if (!observation || typeof observation !== "object") {
-    return null;
-  }
-  if (observation.view && typeof observation.view === "object") {
-    return observation.view;
-  }
-  return observation;
-}
 
 // CR.6 — the `if (lastBaseTiles) return lastBaseTiles` rung is gone; tiles come
 // from this call's payload, this call's observation view, or this call's simConfig.
-function resolveBaseTiles(payload, view, simConfig) {
-  const fromPayload = payload?.baseTiles || payload?.tiles?.baseTiles;
-  if (fromPayload) return fromPayload;
-  if (view?.baseTiles) return view.baseTiles;
-  if (view?.tiles?.baseTiles) return view.tiles.baseTiles;
-  if (view?.tiles?.tiles) return view.tiles.tiles;
-  const config = payload?.simConfig || simConfig;
-  if (config?.layout?.data?.tiles) return config.layout.data.tiles;
-  return null;
-}
 
-function resolveExit(payload, view, baseTiles, simConfigInput) {
-  if (payload?.exit) return payload.exit;
-  if (view?.exit) return view.exit;
-  const simConfig = payload?.simConfig || simConfigInput;
-  if (simConfig?.layout?.data?.exit) return simConfig.layout.data.exit;
-  if (baseTiles) return findExitFromTiles(baseTiles);
-  return null;
-}
 
-function resolveActor(view, actorId, observation) {
-  if (view?.actors && Array.isArray(view.actors)) {
-    const matchId = actorId || observation?.actorId;
-    const selected = matchId ? view.actors.find((actor) => actor?.id === matchId) : view.actors[0];
-    if (selected?.position) {
-      return { id: selected.id, position: selected.position };
-    }
-  }
-  if (view?.actor) {
-    const pos = view.actor.position || (Number.isFinite(view.actor.x) && Number.isFinite(view.actor.y) ? { x: view.actor.x, y: view.actor.y } : null);
-    if (pos) {
-      return { id: view.actor.id || actorId, position: pos };
-    }
-  }
-  if (view?.position) {
-    return { id: actorId || observation?.actorId, position: view.position };
-  }
-  return null;
-}
 
-function resolveActorRecord(view, actorId, observation) {
-  if (view?.actors && Array.isArray(view.actors)) {
-    const matchId = actorId || observation?.actorId;
-    const selected = matchId ? view.actors.find((actor) => actor?.id === matchId) : view.actors[0];
-    if (selected) {
-      return selected;
-    }
-  }
-  if (view?.actor) {
-    return view.actor;
-  }
-  return null;
-}
 
 function resolveConfiguredActor(payload, actorId) {
   const actors = Array.isArray(payload?.initialState?.actors) ? payload.initialState.actors : [];
@@ -454,37 +398,7 @@ function resolveConfiguredActor(payload, actorId) {
   return actors.find((actor) => actor?.id === actorId) || null;
 }
 
-function resolveTileKinds(view, payload) {
-  if (Array.isArray(view?.tiles?.kinds)) return view.tiles.kinds;
-  if (Array.isArray(view?.kinds)) return view.kinds;
-  if (Array.isArray(payload?.tiles?.kinds)) return payload.tiles.kinds;
-  return null;
-}
 
-function buildAdjacentMoveProposals({ actor, tileKinds, baseTiles }) {
-  if (!actor?.position) {
-    return [];
-  }
-  const proposals = [];
-  for (const delta of DEFAULT_DELTAS) {
-    const to = {
-      x: actor.position.x + delta.dx,
-      y: actor.position.y + delta.dy,
-    };
-    if (!isPassable(to, tileKinds, baseTiles)) {
-      continue;
-    }
-    proposals.push({
-      kind: "move",
-      params: {
-        direction: delta.direction,
-        from: actor.position,
-        to,
-      },
-    });
-  }
-  return proposals;
-}
 
 function buildCandidateActionId(proposal, index) {
   if (!proposal || typeof proposal !== "object") {
@@ -724,10 +638,6 @@ function candidateSignature(kind, params) {
   return `${normalizedKind}:${JSON.stringify(isObject(params) ? params : {})}`;
 }
 
-function chebyshevDistance(left, right) {
-  if (!left || !right) return null;
-  return Math.max(Math.abs(right.x - left.x), Math.abs(right.y - left.y));
-}
 
 function nearestHostile(position, visibleActors) {
   if (!position) return null;
@@ -1311,107 +1221,9 @@ function buildRuntimeDecisionEffect({ payload, observation, view, actorId, tick,
   };
 }
 
-function isPassable({ x, y }, tileKinds, baseTiles) {
-  if (tileKinds) {
-    const row = tileKinds[y];
-    if (!Array.isArray(row)) return false;
-    return row[x] === 0;
-  }
-  if (baseTiles) {
-    if (y < 0 || y >= baseTiles.length) return false;
-    const row = String(baseTiles[y]);
-    const cell = row[x];
-    if (!cell) return false;
-    return cell !== "#" && cell !== "B";
-  }
-  return false;
-}
 
-function isDiagonalStepAllowed(current, next, tileKinds, baseTiles) {
-  const dx = next.x - current.x;
-  const dy = next.y - current.y;
-  if (Math.abs(dx) !== 1 || Math.abs(dy) !== 1) {
-    return true;
-  }
-  return isPassable({ x: current.x + dx, y: current.y }, tileKinds, baseTiles)
-    && isPassable({ x: current.x, y: current.y + dy }, tileKinds, baseTiles);
-}
 
-function findPath(start, goal, tileKinds, baseTiles) {
-  if (!start || !goal) return null;
-  if (start.x === goal.x && start.y === goal.y) return [start];
-  const height = tileKinds ? tileKinds.length : baseTiles ? baseTiles.length : 0;
-  const width = tileKinds && Array.isArray(tileKinds[0]) ? tileKinds[0].length : baseTiles && baseTiles[0] ? String(baseTiles[0]).length : 0;
-  if (width === 0 || height === 0) return null;
 
-  const queue = [start];
-  const cameFrom = {};
-  const startKey = `${start.x},${start.y}`;
-  cameFrom[startKey] = null;
-  let head = 0;
-
-  while (head < queue.length) {
-    const current = queue[head];
-    head += 1;
-    if (current.x === goal.x && current.y === goal.y) {
-      const path = [];
-      let key = `${goal.x},${goal.y}`;
-      while (key) {
-        const [x, y] = key.split(",").map((v) => Number(v));
-        path.unshift({ x, y });
-        key = cameFrom[key];
-      }
-      return path;
-    }
-    for (const delta of DEFAULT_DELTAS) {
-      const next = { x: current.x + delta.dx, y: current.y + delta.dy };
-      if (next.x < 0 || next.y < 0 || next.x >= width || next.y >= height) {
-        continue;
-      }
-      const key = `${next.x},${next.y}`;
-      if (Object.prototype.hasOwnProperty.call(cameFrom, key)) {
-        continue;
-      }
-      if (!isPassable(next, tileKinds, baseTiles) || !isDiagonalStepAllowed(current, next, tileKinds, baseTiles)) {
-        continue;
-      }
-      cameFrom[key] = `${current.x},${current.y}`;
-      queue.push(next);
-    }
-  }
-  return null;
-}
-
-function buildMoveProposal({ observation, payload, simConfig }) {
-  const view = resolveObservationView(observation);
-  if (!view) return [];
-  const baseTiles = resolveBaseTiles(payload, view, simConfig);
-  const exit = resolveExit(payload, view, baseTiles, simConfig);
-  const tileKinds = resolveTileKinds(view, payload);
-  const actor = resolveActor(view, payload?.actorId, observation);
-  if (!actor || !actor.position || !exit) return [];
-  const actorRecord = resolveActorRecord(view, payload?.actorId, observation);
-  if (actorRecord?.motivation?.mobility === "stationary") {
-    return [{ kind: "wait", params: { reason: "stationary" } }];
-  }
-  const path = findPath(actor.position, exit, tileKinds, baseTiles);
-  if (!path || path.length < 2) return [];
-  const from = path[0];
-  const to = path[1];
-  const delta = { dx: to.x - from.x, dy: to.y - from.y };
-  const direction = DEFAULT_DELTAS.find((entry) => entry.dx === delta.dx && entry.dy === delta.dy)?.direction;
-  if (!direction) return [];
-  return [
-    {
-      kind: "move",
-      params: {
-        direction,
-        from,
-        to,
-      },
-    },
-  ];
-}
 
 /**
  * PATROLLING: a clockwise circuit of the actor's own room.
@@ -1429,90 +1241,10 @@ function buildMoveProposal({ observation, payload, simConfig }) {
  * would need the runner to thread that state — the problem #162 is still open on. A
  * circuit needs no memory because the ring itself encodes the order.
  */
-function roomPerimeterRing(room, tileKinds, baseTiles) {
-  const x0 = room.x;
-  const y0 = room.y;
-  const x1 = room.x + room.width - 1;
-  const y1 = room.y + room.height - 1;
-  if (x1 <= x0 || y1 <= y0) return [];
-  const ring = [];
-  for (let x = x0; x <= x1; x += 1) ring.push({ x, y: y0 });
-  for (let y = y0 + 1; y <= y1; y += 1) ring.push({ x: x1, y });
-  for (let x = x1 - 1; x >= x0; x -= 1) ring.push({ x, y: y1 });
-  for (let y = y1 - 1; y >= y0 + 1; y -= 1) ring.push({ x: x0, y });
-  // A room may be carved so part of its rectangle is wall. Patrolling the walkable
-  // subset keeps the circuit legal rather than proposing moves core will refuse.
-  return ring.filter((cell) => isPassable(cell, tileKinds, baseTiles));
-}
 
-function roomContaining(position, rooms) {
-  if (!Array.isArray(rooms) || !position) return null;
-  for (const room of rooms) {
-    if (!isObject(room)) continue;
-    const { x, y, width, height } = room;
-    if (![x, y, width, height].every((v) => Number.isInteger(v))) continue;
-    if (position.x >= x && position.x <= x + width - 1
-      && position.y >= y && position.y <= y + height - 1) {
-      return room;
-    }
-  }
-  return null;
-}
 
-function resolveRooms(payload, view, simConfig) {
-  const fromConfig = simConfig?.layout?.data?.rooms;
-  if (Array.isArray(fromConfig)) return fromConfig;
-  const fromPayload = payload?.simConfig?.layout?.data?.rooms;
-  if (Array.isArray(fromPayload)) return fromPayload;
-  const fromView = view?.rooms;
-  return Array.isArray(fromView) ? fromView : [];
-}
 
-function buildPatrolProposals({ observation, payload, simConfig }) {
-  const view = resolveObservationView(observation);
-  if (!view) return [];
-  const actor = resolveActor(view, payload?.actorId, observation);
-  if (!actor?.position) return [];
-  const baseTiles = resolveBaseTiles(payload, view, simConfig);
-  const tileKinds = resolveTileKinds(view, payload);
-  const rooms = resolveRooms(payload, view, simConfig);
-  const room = roomContaining(actor.position, rooms);
-  if (!room) return [];
-  const ring = roomPerimeterRing(room, tileKinds, baseTiles);
-  if (ring.length < 2) return [];
 
-  const at = ring.findIndex((cell) => cell.x === actor.position.x && cell.y === actor.position.y);
-  // On the ring: take the next cell round. Inside the room: walk out to the nearest ring
-  // cell first, so an actor that spawned in the middle joins the patrol instead of
-  // standing still — which would look exactly like the defect this replaces.
-  const target = at >= 0 ? ring[(at + 1) % ring.length] : nearestRingCell(actor.position, ring);
-  if (!target) return [];
-
-  const path = findPath(actor.position, target, tileKinds, baseTiles);
-  if (!path || path.length < 2) return [];
-  const from = path[0];
-  const to = path[1];
-  const direction = DEFAULT_DELTAS.find(
-    (entry) => entry.dx === to.x - from.x && entry.dy === to.y - from.y,
-  )?.direction;
-  if (!direction) return [];
-  return [{ kind: "move", params: { direction, from, to } }];
-}
-
-function nearestRingCell(position, ring) {
-  let best = null;
-  let bestDistance = Infinity;
-  for (const cell of ring) {
-    const distance = chebyshevDistance(position, cell);
-    // Ties break on the ring's own order, which is deterministic — replay compares runs
-    // frame by frame, so "whichever came first" must mean the same thing every run.
-    if (distance < bestDistance) {
-      bestDistance = distance;
-      best = cell;
-    }
-  }
-  return best;
-}
 
 // ── M5: Simple motivation helpers ──────────────────────────────────────────
 
@@ -1617,58 +1349,12 @@ function resolveNearestHostile(view, actorId) {
 // call history, satisfying both the "identical seed -> identical trajectory"
 // and "no shared mutable RNG state" requirements.
 
-function resolveActorRandomSeed(view, actorId, payload, personaSeed) {
-  if (view?.actors && Array.isArray(view.actors)) {
-    const self = view.actors.find((a) => a && a.id === actorId);
-    if (self?.motivation?.seed !== undefined && self.motivation.seed !== null) {
-      return self.motivation.seed;
-    }
-  }
-  const configActors = payload?.initialState?.actors;
-  if (Array.isArray(configActors)) {
-    const configActor = configActors.find((a) => a && a.id === actorId);
-    if (configActor?.motivation?.seed !== undefined && configActor.motivation.seed !== null) {
-      return configActor.motivation.seed;
-    }
-  }
-  if (payload?.seed !== undefined && payload?.seed !== null) return payload.seed;
-  if (personaSeed !== undefined && personaSeed !== null) return personaSeed;
-  return 0;
-}
 
 /** Deterministic 32-bit hash of a seed/actorId/tick tuple (no Math.random, no clock). */
-function hashRandomInputs(seed, actorId, tick) {
-  const text = `${String(seed)}:${String(actorId)}:${String(Number.isFinite(tick) ? tick : 0)}`;
-  let hash = 2166136261;
-  for (let i = 0; i < text.length; i += 1) {
-    hash ^= text.charCodeAt(i);
-    hash = Math.imul(hash, 16777619);
-  }
-  // Final mix (mulberry32-style) so nearby hashes decorrelate before use.
-  hash ^= hash << 13;
-  hash ^= hash >>> 17;
-  hash ^= hash << 5;
-  return hash >>> 0;
-}
 
 /** Pure deterministic RNG value in [0, 1) derived from seed + actorId + tick + salt. */
-function deterministicRandom(seed, actorId, tick, salt = 0) {
-  const hashed = hashRandomInputs(seed, actorId, (Number.isFinite(tick) ? tick : 0) * 2654435761 + salt);
-  return hashed / 4294967296;
-}
 
-function isOccupied(position, view, actorId) {
-  if (!view?.actors || !Array.isArray(view.actors)) return false;
-  return view.actors.some(
-    (other) => other && other.id !== actorId && other.position
-      && other.position.x === position.x && other.position.y === position.y,
-  );
-}
 
-function isReserved(position, reservedTargets) {
-  if (!Array.isArray(reservedTargets) || reservedTargets.length === 0) return false;
-  return reservedTargets.some((target) => target && target.x === position.x && target.y === position.y);
-}
 
 /**
  * Random motivation: choose among legal adjacent walkable, unoccupied tiles
@@ -1759,40 +1445,6 @@ function buildAffinityCastProposal({ actor, affinities: rawAffinities, hostile }
   return null;
 }
 
-function buildRandomMoveProposals({ observation, payload, simConfig, personaSeed }) {
-  const view = resolveObservationView(observation);
-  const actorId = payload?.actorId;
-  const actor = resolveActor(view, actorId, observation);
-  if (!actor?.position) return [{ kind: "wait", params: { reason: "random" } }];
-
-  const baseTiles = resolveBaseTiles(payload, view, simConfig);
-  const tileKinds = resolveTileKinds(view, payload);
-  const reservedTargets = payload?.reservedTargets;
-  const candidates = buildAdjacentMoveProposals({ actor, tileKinds, baseTiles })
-    .filter((proposal) => !isOccupied(proposal.params.to, view, actorId))
-    .filter((proposal) => !isReserved(proposal.params.to, reservedTargets));
-
-  if (candidates.length === 0) {
-    return [{ kind: "wait", params: { reason: "random" } }];
-  }
-
-  const seed = resolveActorRandomSeed(view, actorId, payload, personaSeed);
-  const roll = deterministicRandom(seed, actorId, payload?.tick, 0);
-  const index = Math.floor(roll * candidates.length) % candidates.length;
-  const chosen = candidates[index];
-
-  return [
-    {
-      kind: "move",
-      params: {
-        direction: chosen.params.direction,
-        from: chosen.params.from,
-        to: chosen.params.to,
-        reason: "random",
-      },
-    },
-  ];
-}
 
 function buildMotivatedProposals({ observation, payload, simConfig, personaSeed }) {
   const view = resolveObservationView(observation);
