+++ packages/runtime/src/personas/actor/motivations/friendly.js
@@ -1,0 +1,11 @@
+import { buildMoveProposal } from "../proposal-helpers.js";
+
+export default {
+  kind: "friendly",
+  /**
+   * @returns {Array<{kind: string, params: object}>} candidates, never a decision.
+   */
+  propose({ observation, payload, simConfig, personaSeed, params }) {
+    return buildMoveProposal({ observation, payload, simConfig });
+  },
+};
