+++ packages/runtime/src/personas/actor/motivations/strategy-focused.js
@@ -1,0 +1,10 @@
+export default {
+  kind: "strategy_focused",
+  /**
+   * @returns {Array<{kind: string, params: object}>} candidates, never a decision.
+   * Combat/hold kinds stay empty in MC.1 — routing still owns those branches.
+   */
+  propose({ observation, payload, simConfig, personaSeed, params }) {
+    return [];
+  },
+};
